requirements.txt
streamlit
langchain-core
langchain_community
langchain_groq
# langchain-openai
langchain-google-genai
langchain-tavily
langchain-experimental
langgraph
python-dotenv
fastapi
uvicorn
pydantic
requests
wikipedia
